add strict and option variable environment

+++ types.go
@@ -7,7 +7,7 @@ import (
 var (
 	BOOL   = reflect.TypeOf((*bool)(nil)).Elem()
 	STRING = reflect.TypeOf((*string)(nil)).Elem()
-	INT    = reflect.TypeOf((*int64)(nil)).Elem()
+	INT    = reflect.TypeOf((*Int)(nil)).Elem()
 	FLOAT  = reflect.TypeOf((*float64)(nil)).Elem()
 	ANY    = reflect.TypeOf((*interface{})(nil)).Elem()
 	ATOM   = reflect.TypeOf((*Atom)(nil)).Elem()
